Allow custom implementations of AggregateRootDirectory
